requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
dj-database-url==2.3.0
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.1.3
